+++ packages/demo-next/pages/blocks.tsx
@@ -22,21 +22,19 @@ import { useJsonForm } from 'next-tinacms-json'
 import { ModalProvider, BlockTemplate } from 'tinacms'
 import {
   InlineForm,
-  InlineImageField,
-  InlineTextField,
+  InlineImage,
+  InlineText,
   InlineBlocks,
   BlocksControls,
-  BlockText,
-  BlockImage,
-  BlockTextarea,
   useInlineForm,
+  InlineTextarea,
 } from 'react-tinacms-inline'
 
 /**
  * This is an example page that uses Blocks from Json
  */
 export default function BlocksExample({ jsonFile }) {
-  const [, form] = useJsonForm(jsonFile)
+  const [data, form] = useJsonForm(jsonFile)
 
   return (
     <ModalProvider>
@@ -44,12 +42,15 @@ export default function BlocksExample({ jsonFile }) {
         <EditToggle />
         <DiscardChanges />
         <h1>
-          <InlineTextField name="title" />
-          <InlineImageField
+          <InlineText name="title" />
+          <InlineImage
             name="hero_image"
+            previewSrc={formValues => formValues.hero_image}
             parse={filename => `/images/${filename}`}
             uploadDir={() => '/public/images/'}
-          />
+          >
+            {props => <ChildImage src={data.hero_image} {...props} />}
+          </InlineImage>
         </h1>
         <Wrap>
           <InlineBlocks name="blocks" blocks={PAGE_BUILDER_BLOCKS} />
@@ -57,6 +58,10 @@ export default function BlocksExample({ jsonFile }) {
       </InlineForm>
     </ModalProvider>
   )
+}
+
+function ChildImage(props) {
+  return <img src={props.previewSrc || props.src} />
 }
 
 /**
@@ -114,13 +119,14 @@ function HeroBlock({ index }) {
 function ImageBlock({ index, data }) {
   return (
     <BlocksControls index={index}>
-      <BlockImage
+      <InlineImage
         name="src"
+        previewSrc={formValues => {
+          return formValues.blocks[index].src
+        }}
         parse={filename => `/images/${filename}`}
         uploadDir={() => '/public/images/'}
-      >
-        <img src={data.src} alt={data.alt} />
-      </BlockImage>
+      />
     </BlocksControls>
   )
 }
@@ -139,7 +145,7 @@ const image_template: BlockTemplate = {
 
 // Testing the block styled component override
 
-const StyledBlockText = styled(BlockText)`
+const StyledBlockText = styled(InlineTextarea)`
   color: green;
 `