+++ packages/gatsby-tinacms-git/src/index.ts
@@ -17,3 +17,4 @@ limitations under the License.
 */
 
 export * from './form-actions/delete-action'
+export * from './use-git-form'